requirements.txt
certifi==2020.6.20
Django==3.0.3
djangorestframework==3.11.1
pytz==2020.1
sqlparse==0.3.1
wincertstore==0.2
